MyApps-13: Move get cache book UseCase to submodules.
