+++ domain/di/src/main/java/xyz/torquato/myapps/domain/di/UseCaseImpl.kt
@@ -7,6 +7,8 @@ import dagger.hilt.components.SingletonComponent
 import xyz.torquato.myaps.domain.api.pagging.IPagingRepository
 import xyz.torquato.myaps.domain.api.web.ILocalWebRepository
 import xyz.torquato.myaps.domain.api.web.IQueryRepository
+import xyz.torquato.myaps.domain.api.web.cache.IWebCacheRepository
+import xyz.torquato.myaps.domain.impl.web.GetCacheBooksUseCase
 import xyz.torquato.myaps.domain.impl.web.GetSelectedBookUseCase
 import xyz.torquato.myaps.domain.impl.web.SetQueryUseCase
 import xyz.torquato.myaps.domain.impl.web.content.SetSelectedBookUseCase
@@ -49,5 +51,13 @@ object UseCaseImpl {
         repository: ILocalWebRepository
     ): GetSelectedBookUseCase = GetSelectedBookUseCase(
         repository
+    )
+
+    @Singleton
+    @Provides
+    fun provideGetCacheBooksUseCase(
+        webCacheRepository: IWebCacheRepository
+    ): GetCacheBooksUseCase = GetCacheBooksUseCase(
+        webCacheRepository
     )
 }
